rename base_test to test.py
