+++ app/tests/test_category.py
@@ -1,5 +1,5 @@
 import json
-from app.tests.test_base import BaseTest
+from app.tests.test import BaseTest
 from db_init import delete_category, create_tables
